MIPS: Loongson: CS5536 MFGPT: Fixup of warning
If the HZ is too small, there will be a following warning:

warning: large integer implicitly truncated to unsigned type

This patch fixes it via forcing the constant to be a suitable type.

+++ arch/mips/include/asm/mach-loongson/cs5536/cs5536_mfgpt.h
@@ -24,8 +24,8 @@ static inline void __maybe_unused enable_mfgpt0_counter(void)
 }
 #endif
 
-#define MFGPT_TICK_RATE 14318000
-#define COMPARE  ((MFGPT_TICK_RATE + HZ/2) / HZ)
+#define MFGPT_TICK_RATE 14318000UL
+#define COMPARE  ((u16)(((u32)MFGPT_TICK_RATE + HZ/2) / HZ))
 
 #define MFGPT_BASE	mfgpt_base
 #define MFGPT0_CMP2	(MFGPT_BASE + 2)
